requirements.txt
google-cloud-aiplatform==1.7.0
google-cloud-pipeline-components
kfp
